Rename NumericInput component
