+++ packages/store-ui/src/molecules/NumberInput/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './NumberInput'
+export type { NumberInputProps } from './NumberInput'
