+++ packages/store-ui/src/molecules/NumberInput/stories/NumberInput.mdx
@@ -1,21 +1,21 @@
 import { Canvas, Props, Story, ArgsTable } from '@storybook/addon-docs'
 
-import NumericInput from '../NumericInput'
+import NumberInput from '../NumberInput'
 
-# NumericInput
+# NumberInput
 
 <Canvas>
-  <Story id="molecules-numeric-input--numeric-input" />
+  <Story id="molecules-number-input--number-input" />
 </Canvas>
 
 ## Props
 
-<ArgsTable of={NumericInput} />
+<ArgsTable of={NumberInput} />
 
 ## CSS Selectors
 
 ```css
-[data-store-numeric-input] {
+[data-store-number-input] {
 }
 
 [data-store-button=('dec|inc')] {
