--- packages/store-ui/src/molecules/NumericInput/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './NumericInput'
-export type { NumericInputProps } from './NumericInput'
