+++ themes/theme-b2c-tailwind/src/molecules/index.css
@@ -1,6 +1,6 @@
 @import "./bullets.css";
 @import "./search-input.css";
 @import "./icon-button.css";
-@import "./numeric-input.css";
+@import "./number-input.css";
 @import './price-range.css';
 @import "./carousel.css";
